Put SystemStatusContent component to a separate file.
