+++ public/apps/configuration-react/pages/SystemStatus/components/index.js
@@ -1,1 +1,2 @@
-export { default as UploadLicense } from './UploadLicense'
+export { default as UploadLicense } from './UploadLicense';
+export { default as SystemStatusContent } from './SystemStatusContent';
